Refactor to an object-oriented interface.
The interface has been rewritten to create the interface from five
objects: a spectrometer frame, a file frame, a plotting frame,
a window to create new CSV files and a master window to hold the
three frames in a notebook. The interface looks the same, but should
be easier to keep track of the different variables and functions
required for the program to function. The variables for wavelength,
duration and frequency were also changed to reflect their numeric
nature, and the file opening/creating functions were also changed
to only modify the filename if the file is actually created (as
opposed to a cancellation of the file creation process), or if the
preexisting file passes the checks for the program.