Fix overfrequent call to changewavelength
The changewavelength callback gets called every time the value in the entry
changes, including when the entry is cleared, and after every keystroke.
The current code will catch the TclError from when the entry is cleared,
and the if statement will only allow the wavelength to be changed if three
characters have been entered (a minimum for the spectrometer).