Add widgets for additional fields with new csv
Adjusted the code so that new widgets get created both when an
existing CSV file is opened, and when one is created. The code
for creating the new widgets was modified so that the widgets
are created when csvfile is changed. This means that subsequent
adjustments of the widget code will be coordinated with the
two operations, but does mean that the filename can be changed
to a file that doesn't work for the program.